requirements.txt
requests>=2.31.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv>=1.0.0
pip install dotenv

python -m venv venv
.\venv\Scripts\activate